reorganize  folders, update to run front & backend
