+++ book_webframe/client/public/manifest.json
